# pkg/api/openapi/users.yml
openapi: 3.0.3
info:
  title: Users
  version: 0.1.0


paths:
  /users:
    post:
      summary: Create a new user
      operationId: createUser
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivUser'
        '400':
          $ref: './lib/errs.yml#/components/responses/BadRequest'
        '409':
          $ref: './lib/errs.yml#/components/responses/Conflict'
        default:
          $ref: './lib/errs.yml#/components/responses/InternalError'

  /users/me:
    get:
      summary: Get own private info
      operationId: getMe
      tags: [users]
      security:
        - TrustMeBro: []
        - TelegramMiniApp: []
        - SessionKey: []
      responses:
        '200':
          description: Private user info of the caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivUser'
        default:
          $ref: './lib/errs.yml#/components/responses/InternalError'
    patch:
      security:
        - TrustMeBro: []
        - TelegramMiniApp: []
        - SessionKey: []
      summary: Update own profile
      operationId: updateMe
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated private user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivUser'
        '400':
          $ref: './lib/errs.yml#/components/responses/BadRequest'
        '409':
          $ref: './lib/errs.yml#/components/responses/Conflict'
        default:
          $ref: './lib/errs.yml#/components/responses/InternalError'
    delete:
      security:
        - TrustMeBro: []
        - TelegramMiniApp: []
        - SessionKey: []
      summary: Delete own account
      operationId: deleteMe
      tags: [users]
      responses:
        '204':
          description: Account deleted
        default:
          $ref: './lib/errs.yml#/components/responses/InternalError'

  /users/{userID}:
    parameters:
      - $ref: '#/components/parameters/UserID'
    get:
      security:
        - {}
        - TrustMeBro: []
        - TelegramMiniApp: []
        - SessionKey: []
      summary: Get public user info by ID
      operationId: getUserByID
      tags: [users]
      responses:
        '200':
          description: Public user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PubUser'
        '404':
          $ref: './lib/errs.yml#/components/responses/NotFound'
        default:
          $ref: './lib/errs.yml#/components/responses/InternalError'

  /users/username/{username}:
    parameters:
      - $ref: '#/components/parameters/Username'
    get:
      security:
        - {}
        - TrustMeBro: []
        - TelegramMiniApp: []
        - SessionKey: []
      summary: Get public user info by username
      operationId: getUserByUsername
      tags: [users]
      responses:
        '200':
          description: Public user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PubUser'
        '404':
          $ref: './lib/errs.yml#/components/responses/NotFound'
        default:
          $ref: './lib/errs.yml#/components/responses/InternalError'

  /users/username/{username}/availability:
    parameters:
      - $ref: '#/components/parameters/Username'
    get:
      summary: Check username availability
      operationId: checkUsernameAvailability
      tags: [users]
      responses:
        '200':
          description: Username availability status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsernameAvailability'
        '400':
          $ref: './lib/errs.yml#/components/responses/BadRequest'
        default:
          $ref: './lib/errs.yml#/components/responses/InternalError'


components:
  securitySchemes:
    TrustMeBro:
      $ref: './lib/auth.yml#/components/securitySchemes/TrustMeBro'
    TelegramMiniApp:
      $ref: './lib/auth.yml#/components/securitySchemes/TelegramMiniApp'
    SessionKey:
      $ref: './lib/auth.yml#/components/securitySchemes/SessionKey'

  parameters:
    UserID:
      name: userID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 018e4c71-5f2a-7c3d-b8e9-1a2b3c4d5e6f

    Username:
      name: username
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
      example: anderson

  schemas:
    UUID:
      type: string
      format: uuid
      example: d1b3aef0-b669-471a-80c7-f079f87e8706

    NewUserRequest:
      type: object
      required: [id, username, locale]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        username:
          type: string
          minLength: 1
          maxLength: 64
          example: anderson
        locale:
          type: string
          example: sr-RS
        bio:
          type: string
          maxLength: 512
          example: Go developer. CTF enjoyer.

    UpdateUserRequest:
      type: object
      minProperties: 1
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 64
          example: anderson_new
        locale:
          type: string
          example: en-US
        bio:
          type: string
          maxLength: 512
          example: Updated bio.

    PubUser:
      type: object
      required: [id, username, bio]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        username:
          type: string
          example: anderson
        bio:
          type: string
          example: Go developer. CTF enjoyer.

    UsernameAvailability:
      type: object
      required: [available]
      properties:
        available:
          type: boolean
          example: true

    PrivUser:
      allOf:
        - $ref: '#/components/schemas/PubUser'
        - type: object
          required: [locale, created_at, updated_at]
          properties:
            locale:
              type: string
              example: sr-RS
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time